refactor: 유틸함수들 파일 분리
