+++ src/constants/photos.ts
@@ -1,0 +1,35 @@
+import gatsby from "../assets/images/gatsby.webp";
+import gleaning from "../assets/images/gleaning.webp";
+import jobs from "../assets/images/jobs.webp";
+import lastMeal from "../assets/images/lastMeal.webp";
+import monaLisa from "../assets/images/monaLisa.webp";
+import pearl from "../assets/images/pearl.webp";
+import scream from "../assets/images/scream.webp";
+import starNight from "../assets/images/starNight.webp";
+import venus from "../assets/images/venus.webp";
+import einstein from "../assets/images/einstein.webp";
+import jihyeon from "../assets/images/jihyeon.webp";
+import monroe from "../assets/images/monroe.webp";
+import sunflower from "../assets/images/sunflower.webp";
+import karina from "../assets/images/karina.webp";
+import sonOfHuman from "../assets/images/sonOfHuman.webp";
+import dream from "../assets/images/dream.webp";
+
+export const photos = [
+  gatsby,
+  gleaning,
+  jobs,
+  lastMeal,
+  monaLisa,
+  pearl,
+  scream,
+  starNight,
+  venus,
+  einstein,
+  jihyeon,
+  monroe,
+  sunflower,
+  karina,
+  sonOfHuman,
+  dream,
+];
