+++ src/utils/randomInt.ts
@@ -1,0 +1,3 @@
+export const randomInt = (max: number, min: number = 0) => {
+  return Math.floor(Math.random() * max - min) + min;
+};
