+++ src/utils/triangle.ts
@@ -1,5 +1,5 @@
 import { THUMBNAIL_WIDTH, TRIANGLE_GAP } from "../constants/triangle";
-import { Coordinate } from "./../types/index";
+import { Coordinate } from "../types/index";
 
 export const getCoordinate = (index: number) => {
   const coordinate = {
@@ -12,8 +12,4 @@ export const getCoordinate = (index: number) => {
 
 export const getRgbaPixel = (coordinate: Coordinate) => {
   return Math.floor(coordinate.x + coordinate.y * THUMBNAIL_WIDTH);
-};
-
-export const randomInt = (max: number, min: number = 0) => {
-  return Math.floor(Math.random() * max - min) + min;
 };
